Make all the tests pass using JDK8. Update Gradle.

+++ Java/src/test/java/uk/org/winder/maths/Test_Factorial_JUnit4_QuickCheck_Java.java
@@ -26,38 +26,38 @@ public final class Test_Factorial_JUnit4_QuickCheck_Java {
   @Rule
   public ExpectedException thrown = ExpectedException.none();
 
-  @Property public void iterative_positive(@InRange(min = "1", max = "500") final Integer i) {
+  @Property public void iterative_positive(@InRange(min = "1", max = "500") final Long i) {
     assertEquals(iterative(i), BigInteger.valueOf(i).multiply(iterative(i - 1)));
   }
 
-  @Property public void iterative_negative(@InRange(min = "-500", max = "-1") final Integer i) {
+  @Property public void iterative_negative(@InRange(min = "-500", max = "-1") final Long i) {
     thrown.expect(RuntimeException.class);
     iterative(i);
   }
 
-  @Property public void naïveRecursive_positive(@InRange(min="1", max="500") final Integer i) {
+  @Property public void naïveRecursive_positive(@InRange(min="1", max="500") final Long i) {
     assertEquals(naïveRecursive(i), BigInteger.valueOf(i).multiply(naïveRecursive(i - 1)));
   }
 
-  @Property public void naïveRecursive_negative(@InRange(min="-500", max="-1") final Integer i) {
+  @Property public void naïveRecursive_negative(@InRange(min="-500", max="-1") final Long i) {
     thrown.expect(RuntimeException.class);
     naïveRecursive(i);
   }
 
-  @Property public void tailRecursive_positive(@InRange(min="1", max="500") final Integer i) {
+  @Property public void tailRecursive_positive(@InRange(min="1", max="500") final Long i) {
     assertEquals(tailRecursive(i), BigInteger.valueOf(i).multiply(tailRecursive(i - 1)));
   }
 
-  @Property public void tailRecursive_negative(@InRange(min="-500", max="-1") final Integer i) {
+  @Property public void tailRecursive_negative(@InRange(min="-500", max="-1") final Long i) {
     thrown.expect(RuntimeException.class);
     tailRecursive(i);
   }
 
-  @Property public void reductive_positive(@InRange(min="1", max="500") final Integer i) {
+  @Property public void reductive_positive(@InRange(min="1", max="500") final Long i) {
     assertEquals(reductive(i), BigInteger.valueOf(i).multiply(reductive(i - 1)));
   }
 
-  @Property public void reductive_negative(@InRange(min="-500", max="-1") final Integer i) {
+  @Property public void reductive_negative(@InRange(min="-500", max="-1") final Long i) {
     thrown.expect(RuntimeException.class);
     reductive(i);
   }
